Fix function arg type

+++ include/checkps.h
@@ -132,7 +132,7 @@ typedef struct {
 void func_80050080(void);
 void func_8004FEE8(int param_1);
 void func_8004FD68(int param_1);
-void func_80051908(void* arg0, u8* arg1, s16 arg2);
+void func_80051908(void* arg0, u8* arg1, s32 arg2);
 s32 func_8001687C(unsigned short);
 
 /**
